Get rid of a few warnings that show up at high warning levels.

+++ SimTKcommon/src/Parallel2DExecutor.cpp
@@ -201,7 +201,8 @@ void Parallel2DExecutorImpl::execute(Parallel2DExecutor::Task& task, Parallel2DE
     // Execute the square blocks in a series of passes.
     
     for (int i = 0; i < (int)squares.size(); ++i) {
-        SquareTask square(*this, task, squares[i], rangeType, false, i == squares.size()-1);
+        SquareTask square(*this, task, squares[i], rangeType, false, 
+                          i == (int)squares.size()-1);
         executor->execute(square, squares[i].size());
     }
 }
